Exit 1 if an exception is raised while running PHPMD
Same idea as
https://github.com/codeclimate/codeclimate-phpcodesniffer/pull/38 to
make sure the build errors correctly

+++ engine.php
@@ -36,7 +36,14 @@ $runner->queueDirectory("/code");
 
 $server->process_work(true);
 
-foreach ($server->get_all_results() as $result_file) {
-    echo file_get_contents($result_file);
-    unlink($result_file);
+$results = $server->get_all_results();
+
+// If there is no output from the runner, an exception must have occurred
+foreach ($results as $result_file) {
+    if ($result_file = "cc_exception_thrown") {
+        exit(1);
+    } else {
+        echo file_get_contents($result_file);
+        unlink($result_file);
+    }
 }
